feat(tests): add connection with child components test suite
